Renamed context injection tests to follow integration test naming pattern exclusively
